+++ CHANGELOG.md
@@ -1,10 +1,23 @@
+<a name="unreleased"></a>
+## [Unreleased]
 
-<a name="v0.1.0"></a>
-## v0.1.0
-
-> 2022-10-20
+### Chore
+- format comments
 
 ### Docs
+- update docs
 
-* added docs support
+### Feat
+- general update
 
+### Refactor
+- moved caller from hints file to caller file
+
+
+<a name="v0.1.0"></a>
+## v0.1.0 - 2022-10-20
+### Docs
+- added docs support
+
+
+[Unreleased]: https://github.com/tigorlazuardi/tower/compare/v0.1.0...HEAD